backend/requirements.txt
# ==============================================================================
# IP-KVM Control Plane — Python dependencies
# Pinned to latest available versions with pre-built wheels.
# ==============================================================================

# --- Web Framework ---
fastapi==0.115.12
uvicorn[standard]==0.34.0          # ASGI server with uvloop + httptools

# --- Database (async) ---
sqlalchemy[asyncio]==2.0.48
asyncpg==0.31.0                    # async PostgreSQL driver (wheels for py3.14)
alembic==1.15.1                    # schema migrations

# --- Configuration & Validation ---
pydantic==2.12.5
pydantic-settings==2.13.1           # reads .env / env vars into Settings

# --- Security ---
python-jose[cryptography]==3.3.0   # JWT encode / decode (HS256)
bcrypt==4.3.0
# --- HTTP client (healthcheck poller, SDP relay) ---
httpx==0.28.1

# --- WebSocket (proxy bridge) ---
websockets==14.2

# --- Utilities ---
python-multipart==0.0.20           # required for OAuth2PasswordRequestForm
